Fixing collision with the player
I have made the player a character2D node now so collisions are easy
there is now a table sprite which is sectioned as a static2Dbody.

An instance of the player and the table now get generated via the "main"
script which is attached to the "level 1" scene. Click play on main
screen to start playing with the character.