Added random tale generation in the scene
You can now select the number of tables that you want to be generated
(either in the main.gd code or the navigator on the side) these tables
then get placed randomly on the scene.

There is some logic to ensure that they ar not placed over the player,
nor overlapping each other.

This was basically a test of making instances of other scenes (i,e., the
table and player) in another scene (the level)